+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## master
 
 ## 2.18.3RC1
+* Add option "Exclude bxSlider CSS?"
 * Exit if accessed directly
 * RESOLVE #62 Bootstrap class name hide conflict